多重暗黙変換されていたので消去 内容 : sendEffectTriggerがもともとEffectHandleがint32なので多重なのではないかと思い、消去。

+++ include/rayseer.h
@@ -878,7 +878,7 @@ public:
 
         m_manager->SendTrigger(
             static_cast<Effekseer::Handle>(handle.value),
-            static_cast<int32_t>(index)
+            index
         );
 
         return true;
